[BUGFIX] hidden fields should be rendered in extra section
Currently form.select and form.checkbox ViewHelpers render
hidden fields before the actual input field in order to
force submission of a value.
This leads to the situation that labels wrapped around the
field do not correspond to the actual form field, but to
the hidden field.
This is fixed by letting the form ViewHelper taking care of
rendering the requried hidden fields in the already existing
section on top of the form.

Change-Id: I77cc9061a36d2a2984bf479868b1138c513e47f8
Fixes: #26481
Fixes: #13342
Releases: 1.1, 1.2