[TASK] Standalone View: enable escaping for "html"
Since the introduction of Content Negotiation to FLOW3
(I7d753155aa72f41e3df85b076e8067bcbfe7ac04), the default
Format of the ActionRequest is not "html" but undefined.
The Standalone View relied on the default "html" format.

This change enables the escape interceptor if the format
is *either* "html" or not set at all

Change-Id: I4bfed11713740cea95288a39757d91c9d609a49c
Related: #37561
Releases: 1.1, 1.2